backend-services/auth-service/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pymongo
requests
PyJWT
python-dotenv
